rename antd_registry to antd-registry
